tech(server/component): Move Strategy from server to server-core
