+++ component/src/main/java/com/chutneytesting/component/scenario/domain/ComposableStep.java
@@ -8,6 +8,7 @@ import static java.util.Optional.empty;
 import static java.util.Optional.ofNullable;
 import static java.util.stream.Collectors.toMap;
 
+import com.chutneytesting.server.core.domain.execution.Strategy;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.LinkedHashMap;
