+++ component/src/main/java/com/chutneytesting/component/scenario/infra/OrientComposableStepMapper.java
@@ -2,8 +2,8 @@ package com.chutneytesting.component.scenario.infra;
 
 import com.chutneytesting.component.ComposableIdUtils;
 import com.chutneytesting.component.scenario.domain.ComposableStep;
-import com.chutneytesting.component.scenario.domain.Strategy;
 import com.chutneytesting.component.scenario.infra.wrapper.StepVertex;
+import com.chutneytesting.server.core.domain.execution.Strategy;
 import com.orientechnologies.orient.core.db.ODatabaseSession;
 import com.orientechnologies.orient.core.record.OElement;
 import com.orientechnologies.orient.core.record.OVertex;
