+++ component/src/test/java/com/chutneytesting/component/execution/domain/ComposedTestCasePreProcessorTest.java
@@ -11,10 +11,10 @@ import static org.assertj.core.api.Assertions.entry;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
 
-import com.chutneytesting.server.core.domain.dataset.DataSet;
 import com.chutneytesting.component.dataset.domain.DataSetRepository;
-import com.chutneytesting.component.scenario.domain.Strategy;
+import com.chutneytesting.server.core.domain.dataset.DataSet;
 import com.chutneytesting.server.core.domain.execution.ExecutionRequest;
+import com.chutneytesting.server.core.domain.execution.Strategy;
 import com.chutneytesting.server.core.domain.globalvar.GlobalvarRepository;
 import com.chutneytesting.server.core.domain.scenario.TestCaseMetadataImpl;
 import com.chutneytesting.tests.OrientDatabaseHelperTest;
