+++ component/src/test/java/com/chutneytesting/tests/OrientDatabaseHelperTest.java
@@ -8,7 +8,6 @@ import static org.assertj.core.api.Assertions.assertThat;
 
 import com.chutneytesting.component.scenario.domain.ComposableStep;
 import com.chutneytesting.component.scenario.domain.ComposableStepRepository;
-import com.chutneytesting.component.scenario.domain.Strategy;
 import com.chutneytesting.component.scenario.infra.ExecutableComposedStepMapper;
 import com.chutneytesting.component.scenario.infra.ExecutableComposedTestCaseMapper;
 import com.chutneytesting.component.scenario.infra.OrientComposableStepMapper;
@@ -18,6 +17,7 @@ import com.chutneytesting.component.scenario.infra.orient.OrientConfigurationPro
 import com.chutneytesting.component.scenario.infra.orient.OrientDBManager;
 import com.chutneytesting.component.scenario.infra.orient.changelog.OrientChangelogExecutor;
 import com.chutneytesting.component.scenario.infra.wrapper.StepVertex;
+import com.chutneytesting.server.core.domain.execution.Strategy;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.SerializationFeature;
