+++ server-core/src/main/java/com/chutneytesting/server/core/domain/execution/Strategy.java
@@ -1,4 +1,4 @@
-package com.chutneytesting.scenario.domain.gwt;
+package com.chutneytesting.server.core.domain.execution;
 
 import static java.util.Collections.emptyMap;
 import static java.util.Optional.empty;
@@ -11,6 +11,7 @@ import java.util.Optional;
 public class Strategy {
 
     public static final Optional<Strategy> NONE = empty();
+    public static final Strategy DEFAULT = new Strategy("", emptyMap());
 
     public final String type;
     public final Map<String, Object> parameters;
